Audio update
-> Incorporation of the audio system from the main branch (removal of
singleton)
-> Added new PlaySound able to create and get the music as needed
-> Added the ability to change a specific SFX volume (useful when to
modify one sound used that is too loud or too silent compare to the rest
(due to the file itself))
-> Added errors check up when playing a sound
-> Corrected a warning in sprite.cpp
-> Added Clear() function to the ResourceManager

+++ src/Core/Sprite.cpp
@@ -103,8 +103,8 @@ void Sprite::ApplyTexture(SDL_Renderer* renderer)
 **/
 void Sprite::Scale(float k)
 {
-	dstRect->w *= k;
-	dstRect->h *= k;
+	dstRect->w *= (int) k;
+	dstRect->h *= (int) k;
 }
 
 /**
